feat: [UIE-8621] - IAM RBAC: fix tables sorting (#12092)
* feat: [UIE-8621] - IAM RBAC: fix tables sorting

* clean up for AssignedRoles and AssignedEntities tables

* update the sortic logic to make it case-insensitive

* changesets

* fixed unit test

---------
