+++ packages/api-v4/.changeset/pr-12092-upcoming-features-1745405127553.md
@@ -1,0 +1,5 @@
+---
+"@linode/api-v4": Upcoming Features
+---
+
+IAM RBAC: add new type ([#12092](https://github.com/linode/manager/pull/12092))
