+++ packages/manager/.changeset/pr-12092-upcoming-features-1745405097507.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Upcoming Features
+---
+
+IAM RBAC: Fix the sorting in the AssignedRolesTable and AssignedEntitiesTable ([#12092](https://github.com/linode/manager/pull/12092))
