+++ packages/manager/src/features/IAM/Shared/AssignedRolesTable/ChangeRoleDrawer.tsx
@@ -21,8 +21,8 @@ import {
 import { AssignedPermissionsPanel } from '../AssignedPermissionsPanel/AssignedPermissionsPanel';
 import { getAllRoles, getRoleByName, updateUserRoles } from '../utilities';
 
-import type { EntitiesOption } from '../types';
-import type { DrawerModes, ExtendedRoleMap, RolesType } from '../utilities';
+import type { DrawerModes, EntitiesOption, ExtendedRoleMap } from '../types';
+import type { RolesType } from '../utilities';
 
 interface Props {
   mode: DrawerModes;
