+++ packages/manager/src/features/IAM/Users/UserRoles/AssignedEntities.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import { useCalculateHiddenItems } from '../../hooks/useCalculateHiddenItems';
 
-import type { CombinedEntity, ExtendedRoleMap } from '../../Shared/utilities';
+import type { CombinedEntity, ExtendedRoleMap } from '../../Shared/types';
 import type { AccountAccessRole, EntityAccessRole } from '@linode/api-v4';
 
 interface Props {
@@ -59,7 +59,6 @@ export const AssignedEntities = ({
         <Chip
           data-testid="entities"
           deleteIcon={<CloseIcon data-testid="CloseIcon" />}
-          key={entity.id}
           label={entity.name}
           onDelete={() => onRemoveAssignment(entity, role)}
           sx={{
@@ -90,7 +89,7 @@ export const AssignedEntities = ({
       {numHiddenItems > 0 && (
         <Box
           sx={{
-            aligIitems: 'center',
+            alignItems: 'center',
             backgroundColor:
               theme.name === 'light'
                 ? theme.tokens.color.Ultramarine[20]
